[None][test] Consolidate session_reuse/session_prefetcher shared helpers
Both layers hand a live MpiPoolSession to a test that did not spawn it,
so they share the same invariants — and duplicated the code for them:
the spawn-time env+sys.path snapshot, the GPU-memory settle poll, and
CUDA_VISIBLE_DEVICES resolution. Extract those into
tests/test_common/_session_utils.py; policy (what to do on mismatch or a
still-busy GPU) stays in each layer's thin wrapper.

Side effects of sharing one implementation:
- session_reuse's settle barrier now understands the UUID/MIG forms of
  CUDA_VISIBLE_DEVICES (previously fell back to polling ALL GPUs on
  shared CI nodes) and treats an explicitly empty value as no GPUs.
- SessionReuseCache gains a public is_active() probe, replacing the
  prefetcher's reach into REUSE.enabled — seam-yield coordination now
  goes through a declared API (suspension via private_mpi_session
  deliberately does not affect seam ownership).
